requirements.txt
<<<<<<< HEAD
# Core framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Vector database
qdrant-client==1.7.3

# Redis (Session Memory Store - optional for production)
redis==5.0.1

# Embeddings (ONNX Runtime — no PyTorch dependency)
onnxruntime==1.20.1
tokenizers==0.21.1
numpy>=1.24.0

# Java parsing
tree-sitter==0.21.3
tree-sitter-java==0.21.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.3
requests>=2.31.0

# Configuration
pyyaml==6.0.1

# Tokenization (accurate token counting for Qwen / code)
tiktoken==0.6.0

# Utilities
tenacity==8.2.3
structlog==24.1.0

# LangGraph orchestration
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
=======
# === Cloud VM Server ===

# Core framework
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.6

# LangGraph agent orchestration
langgraph>=0.2
langchain-core>=0.3

# Vector database
qdrant-client>=1.12

# Dense embedding model
sentence-transformers>=3.3

# BM25 sparse vector computation
rank-bm25>=0.2

# vLLM client (OpenAI-compatible)
openai>=1.54

# Environment
python-dotenv>=1.0

# HTTP client
httpx>=0.27

# === Testing ===
pytest>=8.0
pytest-asyncio>=0.23

# === MCP Server (Client Machine) ===

# MCP protocol
mcp>=1.0

# Tree-sitter parsing
tree-sitter>=0.23
tree-sitter-java
tree-sitter-go
tree-sitter-python
tree-sitter-c-sharp
tree-sitter-javascript
tree-sitter-hcl
>>>>>>> feature/new-architecture
